requirements.txt
beautifulsoup4==4.14.3
gspread==6.2.1
gspread_dataframe==4.0.0
pandas==3.0.1
protobuf==7.34.0
pytest==9.0.2
Requests==2.32.5
SQLAlchemy==2.0.48
psycopg2-binary==2.9.9
google-auth==2.29.0
pytest-cov==5.0.0